requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
odmantic==1.0.2
motor==3.7.1
pymongo==4.15.3
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic==2.11.9
PyJWT==2.10.1
bcrypt==5.0.0
email-validator==2.3.0
cloudinary==1.44.1
pillow==12.0.0
requests==2.32.5
huggingface-hub==0.36.0
